+++ src/PermissionManagerServiceProvider.php
@@ -35,17 +35,11 @@ class PermissionManagerServiceProvider extends ServiceProvider
 
         // use the vendor configuration file as fallback
         $this->mergeConfigFrom(
-            __DIR__.'/config/permission.php', 'permission'
-        );
-        $this->mergeConfigFrom(
             __DIR__.'/config/backpack/permissionmanager.php', 'backpack.permissionmanager'
         );
 
         // publish config file
         $this->publishes([__DIR__.'/config' => config_path()], 'config');
-
-        // publish migrations
-        $this->publishes([__DIR__.'/database/migrations' => database_path('migrations')], 'migrations');
 
         // publish translation files
         $this->publishes([__DIR__.'/resources/lang' => resource_path('lang/vendor/backpack')], 'lang');