+++ src/app/Http/Controllers/PermissionCrudController.php
@@ -4,47 +4,33 @@ namespace Backpack\PermissionManager\app\Http\Controllers;
 
 use Backpack\CRUD\app\Http\Controllers\CrudController;
 // VALIDATION
-use Backpack\PermissionManager\app\Http\Requests\RoleCrudRequest as StoreRequest;
-use Backpack\PermissionManager\app\Http\Requests\RoleCrudRequest as UpdateRequest;
+use Backpack\PermissionManager\app\Http\Requests\PermissionCrudRequest as StoreRequest;
+use Backpack\PermissionManager\app\Http\Requests\PermissionCrudRequest as UpdateRequest;
 
 class PermissionCrudController extends CrudController
 {
     public function setup()
     {
-        $role_model = config('permission.models.role');
-        $permission_model = config('permission.models.permission');
+        $role_model = config('backpack.permissionmanager.models.role');
+        $permission_model = config('backpack.permissionmanager.models.permission');
 
-        $this->crud->setModel($role_model);
-        $this->crud->setEntityNameStrings(trans('backpack::permissionmanager.role'), trans('backpack::permissionmanager.roles'));
-        $this->crud->setRoute(config('backpack.base.route_prefix').'/role');
+        $this->crud->setModel($permission_model);
+        $this->crud->setEntityNameStrings(trans('backpack::permissionmanager.permission_singular'), trans('backpack::permissionmanager.permission_plural'));
+        $this->crud->setRoute(backpack_url('permission'));
 
-        $this->crud->setColumns([
-            [
-                'name'  => 'name',
-                'label' => trans('backpack::permissionmanager.name'),
-                'type'  => 'text',
-            ],
-            [
-                'name'  => 'guard_name',
-                'label' => trans('backpack::permissionmanager.guard_type'),
-                'type'  => 'text',
-            ],
-            [
-                'name'  => 'guard_name',
-                'label' => trans('backpack::permissionmanager.guard_type'),
-                'type'  => 'text',
-            ],
-            [
-                // n-n relationship (with pivot table)
-                'label'     => ucfirst(trans('backpack::permissionmanager.permission_plural')),
-                'type'      => 'select_multiple',
-                'name'      => 'permissions', // the method that defines the relationship in your Model
-                'entity'    => 'permissions', // the method that defines the relationship in your Model
-                'attribute' => 'name', // foreign key attribute that is shown to user
-                'model'     => $permission_model, // foreign key model
-                'pivot'     => true, // on create&update, do you need to add/delete pivot table entries?
-            ],
+        $this->crud->addColumn([
+            'name'  => 'name',
+            'label' => trans('backpack::permissionmanager.name'),
+            'type'  => 'text',
         ]);
+
+        if (config('backpack.permissionmanager.multiple_guards')) {
+            $this->crud->addColumn([
+                'name'  => 'guard_name',
+                'label' => trans('backpack::permissionmanager.guard_type'),
+                'type'  => 'text',
+            ]);
+        }
 
         $this->crud->addField([
             'name'  => 'name',
@@ -52,29 +38,22 @@ class PermissionCrudController extends CrudController
             'type'  => 'text',
         ]);
 
-        $this->crud->addField([
-            'name'    => 'guard_name',
-            'label'   => trans('backpack::permissionmanager.guard_type'),
-            'type'    => 'select_from_array',
-            'options' => $this->getGuardTypes(),
-        ]);
-        $this->crud->addField([
-            'label'     => ucfirst(trans('backpack::permissionmanager.permission_plural')),
-            'type'      => 'checklist',
-            'name'      => 'permissions',
-            'entity'    => 'permissions',
-            'attribute' => 'name',
-            'model'     => $permission_model,
-            'pivot'     => true,
-        ]);
+        if (config('backpack.permissionmanager.multiple_guards')) {
+            $this->crud->addField([
+                'name'    => 'guard_name',
+                'label'   => trans('backpack::permissionmanager.guard_type'),
+                'type'    => 'select_from_array',
+                'options' => $this->getGuardTypes(),
+            ]);
+        }
 
-        if (config('backpack.permissionmanager.allow_role_create') == false) {
+        if (config('backpack.permissionmanager.allow_permission_create') == false) {
             $this->crud->denyAccess('create');
         }
-        if (config('backpack.permissionmanager.allow_role_update') == false) {
+        if (config('backpack.permissionmanager.allow_permission_update') == false) {
             $this->crud->denyAccess('update');
         }
-        if (config('backpack.permissionmanager.allow_role_delete') == false) {
+        if (config('backpack.permissionmanager.allow_permission_delete') == false) {
             $this->crud->denyAccess('delete');
         }
     }