+++ src/app/Http/Controllers/RoleCrudController.php
@@ -11,29 +11,36 @@ class RoleCrudController extends CrudController
 {
     public function setup()
     {
-        $role_model = config('permission.models.role');
-        $permission_model = config('permission.models.permission');
+        $role_model = config('backpack.permissionmanager.models.role');
+        $permission_model = config('backpack.permissionmanager.models.permission');
 
         $this->crud->setModel($role_model);
         $this->crud->setEntityNameStrings(trans('backpack::permissionmanager.role'), trans('backpack::permissionmanager.roles'));
-        $this->crud->setRoute(config('backpack.base.route_prefix').'/role');
+        $this->crud->setRoute(backpack_url('role'));
 
-        $this->crud->setColumns([
-            [
-                'name'  => 'name',
-                'label' => trans('backpack::permissionmanager.name'),
+        $this->crud->addColumn([
+            'name'  => 'name',
+            'label' => trans('backpack::permissionmanager.name'),
+            'type'  => 'text',
+        ]);
+
+        if (config('backpack.permissionmanager.multiple_guards')) {
+            $this->crud->addColumn([
+                'name'  => 'guard_name',
+                'label' => trans('backpack::permissionmanager.guard_type'),
                 'type'  => 'text',
-            ],
-            [
-                // n-n relationship (with pivot table)
-                'label'     => ucfirst(trans('backpack::permissionmanager.permission_plural')),
-                'type'      => 'select_multiple',
-                'name'      => 'permissions', // the method that defines the relationship in your Model
-                'entity'    => 'permissions', // the method that defines the relationship in your Model
-                'attribute' => 'name', // foreign key attribute that is shown to user
-                'model'     => $permission_model, // foreign key model
-                'pivot'     => true, // on create&update, do you need to add/delete pivot table entries?
-            ],
+            ]);
+        }
+
+        $this->crud->addColumn([
+            // n-n relationship (with pivot table)
+            'label'     => ucfirst(trans('backpack::permissionmanager.permission_plural')),
+            'type'      => 'select_multiple',
+            'name'      => 'permissions', // the method that defines the relationship in your Model
+            'entity'    => 'permissions', // the method that defines the relationship in your Model
+            'attribute' => 'name', // foreign key attribute that is shown to user
+            'model'     => $permission_model, // foreign key model
+            'pivot'     => true, // on create&update, do you need to add/delete pivot table entries?
         ]);
 
         $this->crud->addField([
@@ -41,6 +48,16 @@ class RoleCrudController extends CrudController
             'label' => trans('backpack::permissionmanager.name'),
             'type'  => 'text',
         ]);
+
+        if (config('backpack.permissionmanager.multiple_guards')) {
+            $this->crud->addField([
+                'name'    => 'guard_name',
+                'label'   => trans('backpack::permissionmanager.guard_type'),
+                'type'    => 'select_from_array',
+                'options' => $this->getGuardTypes(),
+            ]);
+        }
+
         $this->crud->addField([
             'label'     => ucfirst(trans('backpack::permissionmanager.permission_plural')),
             'type'      => 'checklist',
@@ -76,5 +93,23 @@ class RoleCrudController extends CrudController
         \Cache::forget('spatie.permission.cache');
 
         return parent::updateCrud();
+    }
+
+    /*
+     * Get an array list of all available guard types
+     * that have been defined in app/config/auth.php
+     *
+     * @return array
+     **/
+    private function getGuardTypes()
+    {
+        $guards = config('auth.guards');
+
+        $returnable = [];
+        foreach ($guards as $key => $details) {
+            $returnable[$key] = $key;
+        }
+
+        return $returnable;
     }
 }