+++ src/config/backpack/permissionmanager.php
@@ -3,8 +3,19 @@
 return [
 
     /*
-    | Backpack/PermissionManager configs.
+    |--------------------------------------------------------------------------
+    | Models
+    |--------------------------------------------------------------------------
+    |
+    | Models used in the User, Role and Permission CRUDs.
+    |
     */
+
+    'models' => [
+        'user' => App\Models\BackpackUser::class,
+        'permission' => Backpack\PermissionManager\app\Models\Permission::class,
+        'role' => Backpack\PermissionManager\app\Models\Role::class,
+    ],
 
     /*
     |--------------------------------------------------------------------------
@@ -25,5 +36,13 @@ return [
     'allow_role_create'       => true,
     'allow_role_update'       => true,
     'allow_role_delete'       => true,
+
+    /*
+    |--------------------------------------------------------------------------
+    | Multiple-guards functionality
+    |--------------------------------------------------------------------------
+    |
+    */
+    'multiple_guards' => false,
 
 ];